requirements.txt
Flask==3.0.3
flask_wtf==1.2.1
passlib==1.7.4
playwright==1.46.0
Requests==2.32.3
pipenv==2024.0.1
WTForms==3.1.2
gunicorn==23.0.0
google.generativeai
peewee
tidb-vector
